+++ apps/consent-ui/src/components/common/Form/fieldsets/inputs/RadioInput.tsx
@@ -22,7 +22,11 @@
 import { FieldValues, useFormContext } from 'react-hook-form';
 import clsx from 'clsx';
 
-import { FormRadioInputProps } from '../../types';
+import { FormInputProps } from 'src/components/common/Form/types';
+
+type FormRadioInputProps<T extends FieldValues, V extends string> = FormInputProps<T> & {
+	value: V;
+};
 
 const RadioInput = <T extends FieldValues, V extends string>({
 	className,
