+++ apps/consent-ui/src/components/common/Form/fieldsets/inputs/TextInput.tsx
@@ -23,7 +23,16 @@ import { FieldValues, useFormContext } from 'react-hook-form';
 import clsx from 'clsx';
 import { SyntheticEvent } from 'react';
 
-import { FormTextInputProps } from '../../types';
+import {
+	FormFieldSetWithCalloutProps,
+	FormInputProps,
+	FormTextInputType,
+} from 'src/components/common/Form/types';
+
+type FormTextInputProps<T extends FieldValues> = FormInputProps<T> &
+	FormFieldSetWithCalloutProps & {
+		type: FormTextInputType;
+	};
 
 const TextInput = <T extends FieldValues>({
 	className,
